+++ sort/sort_merge.h
@@ -3,6 +3,6 @@
 
 #include <stdio.h>
 
-void sort_merge(void *data, size_t size, size_t element_size, int (*compare)(const void *, const void *));
+void sort_merge(void *data, int size, size_t element_size, int (*compare)(const void *, const void *));
 
 #endif
